Fix error for Controller
