+++ src/main/java/com/tlrh/gestion_tlrh_backend/controllers/CompteController.java
@@ -1,4 +1,4 @@
-package com.tlrh.gestion_tlrh_backend.controller;
+package com.tlrh.gestion_tlrh_backend.controllers;
 
 import com.tlrh.gestion_tlrh_backend.service.CompteService;
 import lombok.AllArgsConstructor;
