+++ src/main/java/com/tlrh/gestion_tlrh_backend/controllers/InitController.java
@@ -1,4 +1,4 @@
-package com.tlrh.gestion_tlrh_backend.controller;
+package com.tlrh.gestion_tlrh_backend.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
