+++ src/main/java/com/tlrh/gestion_tlrh_backend/entity/Certificat.java
@@ -23,4 +23,5 @@ public class Certificat {
     private String nom;
 
     private Date date;
+
 }
