+++ src/main/java/com/tlrh/gestion_tlrh_backend/service/CompteService.java
@@ -5,7 +5,6 @@ import com.tlrh.gestion_tlrh_backend.repositories.CollaborateurRepository;
 import com.tlrh.gestion_tlrh_backend.repositories.CompteRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.Optional;
 
 @Service
